Fix dvc pull issue

+++ starter/main.py
@@ -15,8 +15,8 @@ from model import inference
 if "DYNO" in os.environ and os.path.isdir(".dvc"):
     os.system("dvc config core.no_scm true")
     os.system("dvc config core.hardlink_lock true")
-    # if os.system("dvc pull") != 0:
-    #     exit("dvc pull failed")
+    if os.system("dvc pull") != 0:
+        exit("dvc pull failed")
     os.system("rm -r .dvc .apt/usr/lib/dvc")
 
 app = FastAPI()
